refactor: rename gast files
